+++ RELEASE_NOTES.md
@@ -1,6 +1,12 @@
 ### New in 0.2 (not released yet)
 
 * Breaking: Renamed `ConnectionString` to `MsSqlConnectionString`
+* New: New helpers
+ - `VisualStudioProjectHelpz` Find Visual Studio projects in your project and
+   get information like project location and NuGet packages
+* New: Generate a NuGet package report across all Visual Studio projects
+  enabling you to e.g. test if two different versions of the same package is
+  used
 
 ### New in 0.1.8 (released 2015-09-16)